refactor: split control pruning
